backend/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.22.0
python-multipart==0.0.6
python-jose==3.3.0
google-auth==2.21.0
python-dotenv==1.0.1
Pillow==11.0.0
torch
